Update rubocop gems
This also handles changes to the internal implementation of rubocop
where those changes overlap the internals of the "new" cops that we have
added.  Unfortunately, the the rubocop implementation appears to be
internally inconsistent; some of the cops are using the old mechanism
for autocorrection, and some are using a newer mechanism for
autocorrection.  This commit should bring our changes in line with the
new form of autocorrection, and (we hope) will conform to the internal
expectations of rubocop for the foreseeable future.

Authored-by: Adam Milligan <adam@srp-ok.com>